requirements.txt
streamlit
google-auth
google-cloud-speech
google-generativeai
google-api-python-client